Resolve tests with changes to use the devices. Resolve graphic tests.
https://github.com/fermi-ad/plotting-app/issues/237

+++ lib/widgets/graphic_plot_widget_adapter.dart
@@ -84,7 +84,7 @@ class GraphicPlotWidgetAdapter extends PlotWidgetAdapter {
             // Handle DevScalarArray type
             for (int i = 0; i < deviceValue.value.length; i++) {
               data.add({
-                "Index": i,
+                "Index": i.toDouble(),
                 "Value": deviceValue.value[i],
                 "Channel": channel.name
               });
